requirements.txt
streamlit
geopandas
plotly
fpdf
unidecode
pillow